Pulled out region
